psock decomposed from pcompat
